:art: Update Readme and clean root directory structure
