--- .github/workflows/ci.yml
@@ -1,43 +1,0 @@
-name: CI Core Workflow
-
-# Controls when the action will run.
-on:
-  workflow_run:
-    workflows: ["CI Core Workflow"]
-    types:
-      - completed
-
-
-# A workflow run is made up of one or more jobs that can run sequentially or in parallel
-jobs:
-  # Checks if Build has succeeded.
-  # Returns: `needs.result.outputs.is_success == 'failure' |  `success``
-  publish-buf:
-    runs-on: ubuntu-latest
-    outputs:
-      result: ${{ steps.status.outputs.result }}
-    steps:
-      - name: Determine Status
-        id: status
-        run: echo "::set-output name=result::${{ github.event.workflow_run.conclusion }}"
-
-  # Posts Build Status to Discord
-  notify-result:
-    needs: [status]
-    runs-on: ubuntu-latest
-    steps:
-      - name: Discord Success Notification
-        if: ${{ needs.status.outputs.result == 'success' }}
-        env:
-          DISCORD_WEBHOOK: ${{ secrets.DISCORD_WEBHOOK }}
-        uses: Ilshidur/action-discord@master
-        with:
-          args: 'Core build succeeded!.'
-
-      - name: Discord Failed Notification
-        if: ${{ needs.status.outputs.result != 'success' }}
-        env:
-          DISCORD_WEBHOOK: ${{ secrets.DISCORD_WEBHOOK }}
-        uses: Ilshidur/action-discord@master
-        with:
-          args: 'Failed to build Core.'
