+++ zk/log/logger_null.go
