+++ zk/server.go
@@ -33,15 +33,15 @@ import (
 	"github.com/golang/protobuf/ptypes/empty"
 	grpc_prometheus "github.com/grpc-ecosystem/go-grpc-prometheus"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"github.com/sonr-io/core/zk/config"
 	"github.com/sonr-io/core/crypto/cl"
 	"github.com/sonr-io/core/crypto/ec"
 	"github.com/sonr-io/core/crypto/ecpseudsys"
 	"github.com/sonr-io/core/crypto/ecschnorr"
 	"github.com/sonr-io/core/crypto/pseudsys"
 	"github.com/sonr-io/core/crypto/schnorr"
-	"github.com/sonr-io/core/log"
-	"github.com/sonr-io/core/util"
+	"github.com/sonr-io/core/zk/config"
+	"github.com/sonr-io/core/zk/log"
+	"github.com/sonr-io/core/zk/util"
 	v1 "go.buf.build/grpc/go/sonr-io/core/host/zk/v1"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/codes"
